requirements.txt
GitPython
isstring
public
